contractor-payment-module/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
supabase>=2.4.2
httpx>=0.27.0
python-multipart>=0.0.9
pyzbar==0.1.9
opencv-python-headless>=4.10.0
Pillow>=10.4.0
numpy>=2.0.0
email-validator>=2.1.1
